refactor(tray): split window placement and reopen routing out of lib.rs
`lib.rs` opens with "This file is intentionally thin" and was 2,123 lines. Two
groups came out cleanly, both pure-ish and already unit-tested, so this is a
move rather than a rewrite - no behaviour changes.

- `reopen.rs` (110 lines) - `ReopenTarget`, `is_onboarded`, `reopen_target`, and
  `onboarding_complete`, the gate every dashboard-opening path consults.
- `window_panel.rs` (346 lines) - `tray_anchor_position` / `monitor_work_area`,
  plus the macOS NSPanel helpers (`make_visible_over_fullscreen`, `show_no_focus`,
  `init_as_nspanel`, `install_click_outside_monitor`, `set_process_display_name`).

`onboarding_complete` is re-exported from the crate root on purpose. `tray.rs`
and `commands/system.rs` call it as `crate::onboarding_complete()`, and
`reopen_tests::every_dashboard_path_waits_for_onboarding` asserts on that exact
source string via `include_str!` - re-exporting keeps both the call sites and
the test that guards them unchanged.

Left in place deliberately: `db_encryption_notice_action` (its tests sit beside
it), `tray_debug` (a registered command), and `start_capture` (capture wiring,
its own concern).

lib.rs 2123 -> 1704. All 250 tray tests pass, including the six anchor-position
tests and the three reopen tests under their new module paths.
